Add files via upload
This is the First Semester project of MMCOE FY BTech 2024-25 students: Ayush Sadavarte, Anish Dharmadhikari, Sahil Badve and Piyush Upadhye. This game is a simulation of a history sport called Jousting. This game was also played at the Olympics and has been played since the Roman era. Do cherish your time playing the game and we all hope you enjoy or take help of the code. Thak you :)